move pages components to a new folder
